gpu: nvgpu: fix cert-c violation for irq var
irq number read from dt is unsigned type and it is stored in signed
variable(CERT-INIT31-C). So, make irq number as unsigned.

Jira NVGPU-3438

Reviewed-on: https://git-master.nvidia.com/r/2115393
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/os/linux/module.c
@@ -1362,7 +1362,7 @@ static int gk20a_probe(struct platform_device *dev)
 
 	gk20a->irq_stall = platform_get_irq(dev, 0);
 	gk20a->irq_nonstall = platform_get_irq(dev, 1);
-	if (gk20a->irq_stall < 0 || gk20a->irq_nonstall < 0) {
+	if ((int)gk20a->irq_stall < 0 || (int)gk20a->irq_nonstall < 0) {
 		err = -ENXIO;
 		goto return_err;
 	}
